fix sendfile
Pipelining is complicated.

File to pipe when pipe size is less than file size succeeds with pipe
capacity, but cancels next linked operation. Next file to pipe fails
with EAGAIN if pipe is already full (not drained).
If using hardlink next operation is not canceled, and pipe is drained.
But if that second pipe fails we again don't know how to proceed.

+++ script/bench_strace.sh
@@ -3,27 +3,52 @@ set -e
 
 ulimit -n 65535
 
-conns=1
-secs=1
+conns=100
+secs=2
 
-zig build -Doptimize=ReleaseFast
-strace -c zig-out/bin/httpd --root site/root --cert site/cert_ec --fds 65535 &
+nginx_strace=0
+httpd_strace=1
+httpd_valgrind=0
+httpd_fg=0
 
-script/targets.sh http 8080 localhost testing
-oha -z "$secs"s --no-tui --urls-from-file site/targets-oha -c $conns -w --cacert site/ca/cert.pem
-pkill -USR1 httpd && pkill httpd
+if [ $httpd_fg -eq 1 ]; then
+    zig build -Doptimize=ReleaseFast
+    sudo samply record zig-out/bin/httpd --root site/root --cert site/cert_ec --fds 65535 &
+    script/targets.sh http 8080 localhost testing
+    oha -z "$secs"s --no-tui --urls-from-file site/targets-oha -c $conns -w --cacert site/ca/cert.pem
+    sudo pkill -USR1 httpd && sudo pkill httpd
+fi
 
-exit 0
+if [ $httpd_valgrind -eq 1 ]; then
+    zig build -Doptimize=ReleaseSafe
+    valgrind --tool=callgrind --dump-instr=yes --collect-jumps=yes zig-out/bin/httpd --root site/root --cert site/cert_ec --fds 65535 &
+    script/targets.sh http 8080 localhost testing
+    oha -z "$secs"s --no-tui --urls-from-file site/targets-oha -c $conns -w --cacert site/ca/cert.pem
+    #pkill -USR1 httpd && pkill httpd
 
-nginx -p "$(pwd)" -c script/nginx.conf -g 'daemon off;' &
-nginx_pid=$!
-sleep 0.2
+    pkill -f 'zig-out/bin/httpd'
+fi
 
-worker_pid=$(ps --ppid $nginx_pid -o pid=)
-sudo strace -c -p $worker_pid &
+if [ $httpd_strace -eq 1 ]; then
+    zig build -Doptimize=ReleaseFast
+    strace -c zig-out/bin/httpd --root site/root --cert site/cert_ec --fds 65535 &
+    script/targets.sh http 8080 localhost testing
+    oha -z "$secs"s --no-tui --urls-from-file site/targets-oha -c $conns -w --cacert site/ca/cert.pem
+    pkill -USR1 httpd && pkill httpd
+fi
 
-script/targets.sh http 8081 localhost testing
-oha -z "$secs"s --no-tui --urls-from-file site/targets-oha -c $conns -w --cacert site/ca/cert.pem
-kill $worker_pid
-kill $nginx_pid
-sleep 0.2
+if [ $nginx_strace -eq 1 ]; then
+
+    nginx -p "$(pwd)" -c script/nginx.conf -g 'daemon off;' &
+    nginx_pid=$!
+    sleep 0.2
+
+    worker_pid=$(ps --ppid $nginx_pid -o pid=)
+    sudo strace -c -p $worker_pid &
+
+    script/targets.sh http 8081 localhost testing
+    oha -z "$secs"s --no-tui --urls-from-file site/targets-oha -c $conns -w --cacert site/ca/cert.pem
+    kill $worker_pid
+    kill $nginx_pid
+    sleep 0.2
+fi
